+++ provider/service_resource_test.go
@@ -71,6 +71,8 @@ func TestAccServiceResource_update(t *testing.T) {
 						"firehydrant_service.test_service", "alert_on_add", "true"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rNameUpdated)),
 					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "service_tier", "1"),
@@ -93,8 +95,9 @@ func TestAccServiceResource_update(t *testing.T) {
 	})
 }
 
-func TestAccServiceResource_updateOwnerID(t *testing.T) {
+func TestAccServiceResource_updateLabels(t *testing.T) {
 	rName := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
+	rNameUpdated := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { testFireHydrantIsSetup(t) },
@@ -112,13 +115,66 @@ func TestAccServiceResource_updateOwnerID(t *testing.T) {
 						"firehydrant_service.test_service", "alert_on_add", "true"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rName)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rName)),
 					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "service_tier", "1"),
 				),
 			},
 			{
-				Config: testAccServiceResourceConfig_updateChangeOwnerID(rName),
+				Config: testAccServiceResourceConfig_updateChangeLabels(rNameUpdated),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckServiceResourceExistsWithAttributes_update("firehydrant_service.test_service"),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "alert_on_add", "true"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test2", fmt.Sprintf("test-label2-%s", rNameUpdated)),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "service_tier", "1"),
+				),
+			},
+			{
+				Config: testAccServiceResourceConfig_basic(rNameUpdated),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckServiceResourceExistsWithAttributes_basic("firehydrant_service.test_service"),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "alert_on_add", "false"),
+					// Make sure the labels are not set
+					resource.TestCheckNoResourceAttr(
+						"firehydrant_service.test_service", "labels.test1"),
+					resource.TestCheckNoResourceAttr(
+						"firehydrant_service.test_service", "labels.test2"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "service_tier", "5"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccServiceResource_updateOwnerID(t *testing.T) {
+	rName := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
+	rNameUpdated := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testFireHydrantIsSetup(t) },
+		ProviderFactories: defaultProviderFactories(),
+		CheckDestroy:      testAccCheckServiceResourceDestroy(),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccServiceResourceConfig_update(rName),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					testAccCheckServiceResourceExistsWithAttributes_update("firehydrant_service.test_service"),
 					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
@@ -128,27 +184,63 @@ func TestAccServiceResource_updateOwnerID(t *testing.T) {
 						"firehydrant_service.test_service", "alert_on_add", "true"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rName)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rName)),
 					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "service_tier", "1"),
 				),
 			},
 			{
-				Config: testAccServiceResourceConfig_updateRemoveOwnerID(rName),
+				Config: testAccServiceResourceConfig_update(rNameUpdated),
 				Check: resource.ComposeAggregateTestCheckFunc(
-					testAccCheckServiceResourceExistsWithAttributes_updateRemoveOwnerID("firehydrant_service.test_service"),
+					testAccCheckServiceResourceExistsWithAttributes_update("firehydrant_service.test_service"),
 					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
 					resource.TestCheckResourceAttr(
-						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rName)),
+						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rNameUpdated)),
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "alert_on_add", "true"),
 					resource.TestCheckResourceAttr(
-						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rName)),
+						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rNameUpdated)),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "service_tier", "1"),
+				),
+			},
+			{
+				Config: testAccServiceResourceConfig_updateChangeOwnerID(rNameUpdated),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckServiceResourceExistsWithAttributes_update("firehydrant_service.test_service"),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "alert_on_add", "true"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "description", fmt.Sprintf("test-description-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "labels.test1", fmt.Sprintf("test-label1-%s", rNameUpdated)),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "owner_id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "service_tier", "1"),
+				),
+			},
+			{
+				Config: testAccServiceResourceConfig_basic(rNameUpdated),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckServiceResourceExistsWithAttributes_basic("firehydrant_service.test_service"),
+					resource.TestCheckResourceAttrSet("firehydrant_service.test_service", "id"),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "name", fmt.Sprintf("test-service-%s", rNameUpdated)),
+					resource.TestCheckResourceAttr(
+						"firehydrant_service.test_service", "alert_on_add", "false"),
 					// Make sure owner_id is not set
 					resource.TestCheckResourceAttr(
 						"firehydrant_service.test_service", "owner_id", ""),
 					resource.TestCheckResourceAttr(
-						"firehydrant_service.test_service", "service_tier", "1"),
+						"firehydrant_service.test_service", "service_tier", "5"),
 				),
 			},
 		},
@@ -209,9 +301,9 @@ func testAccCheckServiceResourceExistsWithAttributes_basic(resourceName string) 
 			return fmt.Errorf("Unexpected description. Expected no description, got: %s", serviceResponse.Description)
 		}
 
-		//if !reflect.DeepEqual(serviceResponse.Labels, []string{}) {
-		//	return fmt.Errorf("Bad labels: %v", serviceResponse.Labels)
-		//}
+		if len(serviceResponse.Labels) != 0 {
+			return fmt.Errorf("Unexpected number of labels. Expected no labels, got: %v", len(serviceResponse.Labels))
+		}
 
 		if serviceResponse.Owner != nil {
 			return fmt.Errorf("Unexpected owner. Expected no owner ID, got: %s", serviceResponse.Owner.ID)
@@ -261,9 +353,16 @@ func testAccCheckServiceResourceExistsWithAttributes_update(resourceName string)
 			return fmt.Errorf("Unexpected description. Expected: %s, got: %s", expected, got)
 		}
 
-		//if !reflect.DeepEqual(serviceResponse.Labels, []string{}) {
-		//	return fmt.Errorf("Bad labels: %v", serviceResponse.Labels)
-		//}
+		if len(serviceResponse.Labels) == 0 {
+			return fmt.Errorf("Unexpected number of labels. Expected at least 1 label, got: %v", len(serviceResponse.Labels))
+		}
+
+		for labelKey, labelValue := range serviceResponse.Labels {
+			key := fmt.Sprintf("labels.%s", labelKey)
+			if serviceResource.Primary.Attributes[key] != labelValue {
+				return fmt.Errorf("Unexpected label. Expected %s:%s, got: %s:%s", labelKey, labelValue, labelKey, serviceResource.Primary.Attributes[key])
+			}
+		}
 
 		if serviceResponse.Owner == nil {
 			return fmt.Errorf("Unexpected owner. Expected owner to be set.")
@@ -271,58 +370,6 @@ func testAccCheckServiceResourceExistsWithAttributes_update(resourceName string)
 		expected, got = serviceResource.Primary.Attributes["owner_id"], serviceResponse.Owner.ID
 		if expected != got {
 			return fmt.Errorf("Unexpected owner ID. Expected:%s, got: %s", expected, got)
-		}
-
-		expected, got = serviceResource.Primary.Attributes["service_tier"], fmt.Sprintf("%d", serviceResponse.ServiceTier)
-		if expected != got {
-			return fmt.Errorf("Unexpected service_tier. Expected: %s, got: %s", expected, got)
-		}
-
-		return nil
-	}
-}
-
-func testAccCheckServiceResourceExistsWithAttributes_updateRemoveOwnerID(resourceName string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		serviceResource, ok := s.RootModule().Resources[resourceName]
-		if !ok {
-			return fmt.Errorf("Not found: %s", resourceName)
-		}
-		if serviceResource.Primary.ID == "" {
-			return fmt.Errorf("No ID is set")
-		}
-
-		client, err := firehydrant.NewRestClient(os.Getenv("FIREHYDRANT_API_KEY"))
-		if err != nil {
-			return err
-		}
-
-		serviceResponse, err := client.Services().Get(context.TODO(), serviceResource.Primary.ID)
-		if err != nil {
-			return err
-		}
-
-		expected, got := serviceResource.Primary.Attributes["name"], serviceResponse.Name
-		if expected != got {
-			return fmt.Errorf("Unexpected name. Expected: %s, got: %s", expected, got)
-		}
-
-		expected, got = serviceResource.Primary.Attributes["alert_on_add"], fmt.Sprintf("%t", serviceResponse.AlertOnAdd)
-		if expected != got {
-			return fmt.Errorf("Unexpected alert_on_add. Expected: %s, got: %s", expected, got)
-		}
-
-		expected, got = serviceResource.Primary.Attributes["description"], serviceResponse.Description
-		if expected != got {
-			return fmt.Errorf("Unexpected description. Expected: %s, got: %s", expected, got)
-		}
-
-		//if !reflect.DeepEqual(serviceResponse.Labels, []string{}) {
-		//	return fmt.Errorf("Bad labels: %v", serviceResponse.Labels)
-		//}
-
-		if serviceResponse.Owner != nil {
-			return fmt.Errorf("Unexpected owner. Expected owner to not be set, got: %s.", serviceResponse.Owner.ID)
 		}
 
 		expected, got = serviceResource.Primary.Attributes["service_tier"], fmt.Sprintf("%d", serviceResponse.ServiceTier)
@@ -377,9 +424,31 @@ resource "firehydrant_service" "test_service" {
   name         = "test-service-%s"
   alert_on_add = true
   description  = "test-description-%s"
+  labels = {
+    test1 = "test-label1-%s",
+  }
   owner_id     = firehydrant_team.test_team1.id
   service_tier = "1"
-}`, rName, rName, rName)
+}`, rName, rName, rName, rName)
+}
+
+func testAccServiceResourceConfig_updateChangeLabels(rName string) string {
+	return fmt.Sprintf(`
+resource "firehydrant_team" "test_team1" {
+  name = "test-team1-%s"
+}
+
+resource "firehydrant_service" "test_service" {
+  name         = "test-service-%s"
+  alert_on_add = true
+  description  = "test-description-%s"
+  labels = {
+    test1 = "test-label1-%s",
+    test2 = "test-label2-%s"
+  }
+  owner_id     = firehydrant_team.test_team1.id
+  service_tier = "1"
+}`, rName, rName, rName, rName, rName)
 }
 
 func testAccServiceResourceConfig_updateChangeOwnerID(rName string) string {
@@ -396,17 +465,10 @@ resource "firehydrant_service" "test_service" {
   name         = "test-service-%s"
   alert_on_add = true
   description  = "test-description-%s"
+  labels = {
+    test1 = "test-label1-%s",
+  }
   owner_id     = firehydrant_team.test_team2.id
   service_tier = "1"
-}`, rName, rName, rName, rName)
-}
-
-func testAccServiceResourceConfig_updateRemoveOwnerID(rName string) string {
-	return fmt.Sprintf(`
-resource "firehydrant_service" "test_service" {
-  name         = "test-service-%s"
-  alert_on_add = true
-  description  = "test-description-%s"
-  service_tier = "1"
-}`, rName, rName)
+}`, rName, rName, rName, rName, rName)
 }