+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 BUG FIXES:
 
 * resource/service: Fixed bug that prevented the `description` attribute from being unset ([#51](https://github.com/firehydrant/terraform-provider-firehydrant/pull/51))
+* resource/service: Fixed bug that prevented `labels` from being removed from services ([#52](https://github.com/firehydrant/terraform-provider-firehydrant/pull/52))
 
 ENHANCEMENTS: